+++ BMC-ENGINE/classes/Vector2.h
@@ -1,0 +1,10 @@
+#pragma once
+
+class Vector2 {
+public:
+	int x;
+	int y;
+	
+	Vector2(int x, int y);
+	Vector2();
+};
