--- BMC-ENGINE/classes/Vector2.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-class Vector2 {
-public:
-	int x;
-	int y;
-	
-	Vector2(int x, int y);
-
-	Vector2 zero = new Vector2(0, 0);
-};
-
-Vector2::Vector2(int x, int y)
-{
-	this->x = x;
-	this->y = y;
-}
