+++ BMC-ENGINE/main.cpp
@@ -1,9 +1,5 @@
 ﻿#include "main.h"
 
-#if WIN32
-#include <Windows.h>
-#endif
-
 using namespace std;
 
 bool checkArg(int argc, char* argv[], std::string arg) {
@@ -17,15 +13,15 @@ bool checkArg(int argc, char* argv[], std::string arg) {
 
 int main(int argc, char* argv[])
 {
-	Window window("BMC Engine");
-	window.initImGui();
+	Window* window = new Window("BMC Engine");
+	window->initImGui();
 
 	// OpenGL!
 	Shader defaultShader("./resources/shaders/default/vertex.glsl", "./resources/shaders/default/fragment.glsl");
 
 	Styler::setToDefault();
 
-	DebugUtil debugger(window);
+	DebugUtil debugger(window->getWindow());
 
 	bool useDebugger = checkArg(argc, argv, "--build");
 	
@@ -41,7 +37,7 @@ int main(int argc, char* argv[])
 	
 
 	// main loop
-	while (!glfwWindowShouldClose(window))
+	while (!(window->shouldClose()))
 	{
 		glClearColor(0.6f, 0.5f, 0.8f, 1.0f); 
 		glClear(GL_COLOR_BUFFER_BIT);
@@ -51,13 +47,10 @@ int main(int argc, char* argv[])
 		if(useDebugger)
 			debugger.draw();
 
-		glfwSwapBuffers(window);
+		glfwSwapBuffers(window->getWindow());
 	}
 
-	glfwDestroyWindow(window);
-	glfwTerminate();
-
-	stbi_image_free(icon_image);
+	delete window;
 
 	return 0;
 }
