+++ BMC-ENGINE/window/Window.h
@@ -1,13 +1,19 @@
 #pragma once
 
+//#define STB_IMAGE_IMPLEMENTATION
+
+//#include "stb_image.h"
+
 #include "glad/glad.h"
 #include "GLFW/glfw3.h"
 
 #include <iostream>
 
-#include "stb_image.h"
+#include "Vector2.h"
 
-#include "Vector2.hpp"
+#include "imgui.h"
+#include "imgui_impl_glfw.h"
+#include "imgui_impl_opengl3.h"
 
 class Window {
 public:
@@ -24,10 +30,12 @@ public:
 	void setSize(Vector2 size);
 	void initImGui();
 
+	GLFWwindow* getWindow();
+
 private:
 	void _create(const char* title, int width, int height);
 	void _setSize(int width, int height);
-	void _framebuffer_size_callback(GLFWwindow* window, int width, int height);
+	static void _framebuffer_size_callback(GLFWwindow* window, int width, int height);
 	
 	GLFWwindow* window;
 };
